Rigid hydro geometries can be defined in-memory (#22681)
There was a legacy exception if an in-memory obj was provided as a
rigid hydroelastic object. The rest of the code is sufficient for reading
the mesh from in-memory, so we simply remove the limitation.

The limitation was introduced early in the PR train for in-memory meshes
but never got cleaned up at the end of the train.