Always build Python for wheels (#22745)
Change wheel builds to always build Python from source. This provides
better insulation against what Python versions are or are not packaged
by the build host. Move dependencies for building Python to the base
package list, which allows us to remove logic to install them from
build-python.sh. Improve reporting if the Python source tarball doesn't
match the expected hash. Tweak provision-base.sh to "allow" blank lines
in the package list. (In practice, it seems they were harmless anyway,
but we now explicitly ignore them, as well as comments.)

This does not, however, remove support for using the distro-provided
Python, as we may or may not stick with "always build our own".

While we're at it, also upgrade to the versions of Python shipped with
manylinux 2025.01.05-1.