Introduce 3D scale in mesh processing algorithms (#22772)
This includes parsing mesh into Drake constructs and computing convex
hulls.

When parsing mesh files (e.g., obj, vtk, etc.) we explicitly handle the
case where the scale includes mirroring which may cause the mesh windings
to be wound incorrectly.

This also attempts to unify the internal APIs with regards to what scale
factors are accepted. We accept any scale factor -- even zero. Because
these are internal, if some limit needs to be placed by some application,
we rely on the application to enforce that constraint.

NOTE TO RELEASE ENGINEER:

This is one of multiple PRs that enable non-uniform scaling.  For the release, they should all be lumped together. The total feature description is:

Enable non-uniform scaling for Mesh and Convex geometries.