Update deformable contact point sampling (#22917)
Prior to this commit, when sampling contact points between deformable
and rigid geometries, the contact points are sampled on the rigid
geometry's surface and are equipped with an approximate signed distance
to the deformable geometry's surface.

This PR "reverts" the role of rigid and deformable geometries. Now, the
sample contact points lie on the surface of the deformable geometry and
are equipped with the hydroelastic pressure inherited from the
hydroelastic compliant geometry of the rigid body.

Implications:
- Deformable geometries now only interact with geometries of rigid bodies
  that have a hydroelastic modulus.
- The contact surface change will inevitably lead to change in dynamics.
- In practice, this often significantly reduce the number of contact
  points in deformable vs. rigid contact, and thus boosting sim speed.
- In practice, this mitigates a "sticking" artifact when deformable
  geometry is in contact with a rigid geometry with sharp features.
  However, it introduces a new potential mode of artifact when the
  an inverted deformable element is in contact.