[solvers] Add ProjectedGradientDescentSolver
Implement ProjectedGradientDescentSolver, a class which implements a
first-order method to solve general nonlinear optimization problems.

The specific method is projected gradient descent with backtracking line
search (using the Armijo condition). The method is particularly
effective in cases where the feasible set is convex, as it guarantees a
feasible solution. Furthermore, if the objective function has a
Lipschitz gradient, it will converge to a local optimum.

The user can specify their own gradient and/or projection functions if
more efficient implementations are known.